% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hcru_cohorts.R
\name{compare_hcru_cohorts}
\alias{compare_hcru_cohorts}
\title{Compare HCRU Between Cohorts}
\usage{
compare_hcru_cohorts(hcru_summary, cohort)
}
\arguments{
\item{hcru_summary}{Output of estimate_hcru()}

\item{cohort}{Metadata with person_id and cohort_id}
}
\value{
Group-level summary per domain and time window
}
\description{
Compare HCRU Between Cohorts
}
