% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hcru.R
\name{estimate_hcru}
\alias{estimate_hcru}
\title{Estimate HCRU Events and Costs}
\usage{
estimate_hcru(cohort, hcru, pre_days = 180, post_days = 365)
}
\arguments{
\item{cohort}{Dataframe with person_id and index_date}

\item{hcru}{Dataframe with person_id, event_date, domain, and cost}

\item{pre_days}{Number of days before index (default 180)}

\item{post_days}{Number of days after index (default 365)}
}
\value{
Dataframe with counts and costs per patient per domain and time window
}
\description{
Estimate HCRU Events and Costs
}
