% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcru_trends.R
\name{plot_event_counts}
\alias{plot_event_counts}
\title{plot_event_counts}
\usage{
plot_event_counts(
  summary_df,
  x_var = "period",
  y_var = "Cost",
  cohort_col = "cohort",
  facet_Var = "care_setting",
  title = "Average total cost by domain and cohort",
  x_lable = "Healthcare Setting (Domain)",
  y_lable = "Average total cost",
  fill_lable = "Cohort"
)
}
\arguments{
\item{hcru_group_summary}{Output from estimate_hcru()}

\item{time_window}{}
}
\value{
ggplot object
}
\description{
This plot provide the visualization of the events of the settings/domains
grouped by cohort and time window.
}
\details{
Plot HCRU Event Summary
}
