% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcru_trends.R
\name{plot_hcru}
\alias{plot_hcru}
\title{Plot HCRU Summary}
\usage{
plot_hcru(hcru_group_summary, metric = "mean_count")
}
\arguments{
\item{hcru_group_summary}{Output from compare_hcru_cohorts()}

\item{metric}{Either 'mean_count' or 'mean_cost'}
}
\value{
ggplot object
}
\description{
Plot HCRU Summary
}
