% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hcru.R
\name{estimate_hcru}
\alias{estimate_hcru}
\title{estimate_hcru}
\usage{
estimate_hcru(
  data,
  cohort_col = "cohort",
  patient_id_col = "patient_id",
  admit_col = "admission_date",
  discharge_col = "discharge_date",
  index_col = "index_date",
  visit_col = "visit_date",
  encounter_id_col = "encounter_id",
  setting_col = "care_setting",
  cost_col = "cost_usd",
  readmission_col = "readmission",
  time_window_col = "period",
  los_col = "length_of_stay",
  custom_var_list = NULL,
  pre_days = 180,
  post_days = 365,
  readmission_days_rule = 30,
  group_var = "cohort",
  test = NULL,
  gt_output = FALSE
)
}
\arguments{
\item{data}{A dataframe specifying the health care details}

\item{cohort_col}{A character specifying the name of the cohort column}

\item{patient_id_col}{A character specifying the name of the patient
identifier column}

\item{admit_col}{A character specifying the name of the date of admission
column}

\item{discharge_col}{A character specifying the name of the date of
discharge column}

\item{index_col}{A character specifying the name of the index date or
diagnosis column}

\item{visit_col}{A character specifying the name of the date of
visit/claim column}

\item{encounter_id_col}{A character specifying the name of the
encounter/claim column}

\item{setting_col}{A character specifying the name of the HCRU
setting column e.g. IP, ED, OP, etc.}

\item{cost_col}{A character specifying the name of cost column}

\item{readmission_col}{A character specifying the name of readmission column}

\item{time_window_col}{A character specifying the name of time window column}

\item{los_col}{A character specifying the name of length of stay column}

\item{custom_var_list}{A character vector provides the list of additional
columns}

\item{pre_days}{Number of days before index (default 180 days)}

\item{post_days}{Number of days after index (default 365 days)}

\item{readmission_days_rule}{Rule for how many days can be permissible to
define readmission criteria in AP setting (default 30 days)}

\item{group_var}{A character specifying the name of grouping column}

\item{test}{Optional named list of statistical tests
(e.g. age ~ "wilcox.test").}

\item{gt_output}{A logical values specifying that whether gtsummary output
should be generated or not}
}
\value{
dataframe with HCRU estimates.
}
\description{
This function calculates an estimates of heath care resource utilization
(HCRU) for a given electronic health record data for a given set of settings
e.g. IP, OP, ED/ER, etc.
}
\examples{
estimate_hcru(data = hcru_sample_data)
}
