% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcru_trends.R
\name{plot_hcru}
\alias{plot_hcru}
\title{plot_hcru}
\usage{
plot_hcru(
  summary_df,
  x_var = "time_window",
  y_var = "Cost",
  cohort_col = "cohort",
  facet_var = "care_setting",
  facet_var_n = 3,
  title = "Average total cost by domain and cohort",
  x_label = "Healthcare Setting (Domain)",
  y_label = "Average total cost",
  fill_label = "Cohort"
)
}
\arguments{
\item{summary_df}{Output from estimate_hcru()}

\item{x_var}{A character specifying column name to be plotted on x-axis}

\item{y_var}{A character specifying column name to be plotted on y-axis}

\item{cohort_col}{A character}

\item{facet_var}{A character specifying column name to generate the plot
output in columns}

\item{facet_var_n}{A numeric specifying number of columns for facet output}

\item{title}{A character}

\item{x_label}{A character}

\item{y_label}{A character}

\item{fill_label}{A character}
}
\value{
ggplot object
}
\description{
This function provides the visualization of the events of the settings
grouped by cohort and time window.
}
\details{
Plot HCRU Event Summary
}
