% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hcru_trends.R
\name{plot_hcru}
\alias{plot_hcru}
\title{plot_hcru}
\usage{
plot_hcru(
  summary_df,
  x_var = "period",
  y_var = "Cost",
  cohort_col = "cohort",
  facet_var = "care_setting",
  facet_var_n = 3,
  title = "Average total cost by domain and cohort",
  x_lable = "Healthcare Setting (Domain)",
  y_lable = "Average total cost",
  fill_lable = "Cohort"
)
}
\arguments{
\item{summary_df}{Output from estimate_hcru()}

\item{x_var}{A character specifying column name to be plotted on x-axis}

\item{y_var}{A character specifying column name to be plotted on y-axis}

\item{cohort_col}{A character}

\item{facet_var}{A character specifying column name to generate the plot
output in columns}

\item{facet_var_n}{A character specifying column name to generate the plot
output in n number of columns}

\item{title}{A character}

\item{x_lable}{A character}

\item{y_lable}{A character}

\item{fill_lable}{A character}
}
\value{
ggplot object
}
\description{
This function provide the visualization of the events of the settings/domains
grouped by cohort and time window.
}
\details{
Plot HCRU Event Summary
}
