% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_descriptives}
\alias{summarize_descriptives}
\title{Generate Detailed Descriptive Statistics}
\usage{
summarize_descriptives(
  data,
  patient_id_col = "patient_id",
  setting_col = "care_setting",
  cohort_col = "cohort",
  encounter_id_col = "encounter_id",
  cost_col = "cost_usd",
  los_col = "length_of_stay",
  readmission_col = "readmission",
  time_window_col = "period"
)
}
\arguments{
\item{data}{A dataframe with variables to summarize.}

\item{patient_id_col}{A character specifying the name of patient
identifier column}

\item{setting_col}{A character specifying the name of HRCU setting column}

\item{cohort_col}{A character specifying the name of cohort column}

\item{encounter_id_col}{A character specifying the name of encounter/claim
column}

\item{cost_col}{A character specifying the name of cost column}

\item{los_col}{A character specifying the name of length of stay column}

\item{readmission_col}{A character specifying the name of readmission column}

\item{time_window_col}{A character specifying the name of time window column}
}
\value{
A table object
}
\description{
Generate Detailed Descriptive Statistics
}
