% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_descriptives_gt}
\alias{summarize_descriptives_gt}
\title{Generate Detailed Descriptive Statistics with Custom P-Value Tests}
\usage{
summarize_descriptives_gt(data, var_list = NULL, group_var = NULL, test = NULL)
}
\arguments{
\item{data}{A dataframe with variables to summarize.}

\item{var_list}{Optional quoted variable list (e.g. care_setting).}

\item{group_var}{Optional quoted grouping variable (e.g. cohort).}

\item{test}{Optional named list of statistical tests
(e.g. age ~ "wilcox.test").}
}
\value{
A gtsummary table object
}
\description{
Generate Detailed Descriptive Statistics with Custom P-Value Tests
}
